somthing style change

+++ src/Pages/Admin/UpdadeService/UpdateService.js
@@ -1,6 +1,5 @@
 import axios from 'axios';
 import React, { useEffect, useState } from 'react';
-
 import UpdateServices from './UpdateServices';
 
 const UpdateService = () => {
@@ -13,24 +12,27 @@ const UpdateService = () => {
     }, [setServicesUpdate])
 
     const handleDelete = (id) => {
-        axios.delete(`https://arcane-sierra-37156.herokuapp.com/updateDelete/${id}`)
-            .then(res => {
-                if (res.data.deletedCount) {
-                    fetch('https://arcane-sierra-37156.herokuapp.com/allServices')
-                        .then(res => res.json())
-                        .then(data => setServicesUpdate(data))
-                }
-            })
+        window.confirm("Are you sure you wish to delete this item?") &&
+            axios.delete(`https://arcane-sierra-37156.herokuapp.com/updateDelete/${id}`)
+                .then(res => {
+                    if (res.data.deletedCount) {
+                        fetch('https://arcane-sierra-37156.herokuapp.com/allServices')
+                            .then(res => res.json())
+                            .then(data => setServicesUpdate(data))
+                    }
+                })
 
     }
 
     return (
         <div className='container pb-11'>
             <h1 className='text-yellow-400 fw-bold underline text-center py-11'>All Update Services</h1>
-            <div className='row g-3'>
+            <div className='row row-cols-1 row-cols-md-4 g-4'>
                 {
                     servicesUpdate?.map(serviceUpdate => <UpdateServices key={serviceUpdate?._id} handleDelete={handleDelete} serviceUpdate={serviceUpdate}></UpdateServices>)
                 }
+
+
             </div>
         </div>
     );
